Warnings cleanup

+++ src/Framebuffer.cpp
@@ -104,7 +104,7 @@ void Framebuffer::destroy() {
 	glDeleteFramebuffers(1, &fbo);
 	glDeleteRenderbuffers(1, &rbo);
 
-	for (int i = 0; i < attachments.size(); i++) {
+	for (size_t i = 0; i < attachments.size(); i++) {
 		attachments[i].destroy();
 	}
